Remove redundant functions
Removes the `LinkLft` and `LinkRgt` functions in favor of a single
`LinkHorz` function which behaves identically to the old `LinkLft`
function. Since the lists are circular, only a single function is
needed for both scenarios.

    LinkRgt(a, b)       // same thing
    LinkLft(a, b.Rgt()) // same thing

Adding `LinkVert` as an exact analog to `LinkHorz` works, though
the tests have to change to reflect the business logic on Head and
Column Nodes.