Redesign `AddRow()` to simplify use
Redesigns the `AddRow()` method to accept an array of column names
instead of using the callback closure. This simplifies use and has
the potential to reduce the overall time and memory required to
pass a partial solution. Instead of having to iterate through
all constraints on the consumer side, we only need to pass the set
of constraints we satisfy. This works because we have enforced
unique column names.